Move emoji filter to own file
