+++ lib/reveal-ck/builders.rb
@@ -8,3 +8,4 @@ require_relative 'builders/create_slides_html'
 require_relative 'builders/create_index_html'
 require_relative 'builders/user_files'
 require_relative 'builders/index_html'
+require_relative 'builders/reveal_ck_emoji_filter'
